+++ src/main/java/ruiseki/omoshiroikamo/module/multiblock/common/block/solarArray/TESolarArray.java
@@ -67,8 +67,8 @@ public abstract class TESolarArray extends AbstractMBModifierTE implements IOKEn
     }
 
     @Override
-    public float getSpeedMultiplier() {
-        return 1.0F;
+    public double getSpeedMultiplier() {
+        return 1.0;
     }
 
     @Override
